requirements.txt
gradio>=2.9.0,<2.10.0
torch>=1.10.0
git+https://github.com/openai/CLIP.git
Pillow
matplotlib
numpy
